# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoloQuiz API",
        "description": "API endpoints for EcoloQuiz - Quiz and Player Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/email-templates": {
            "get": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Get all email templates",
                "operationId": "eb6c27df7218121b5b87ad98b6475463",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of email templates"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Create a new email template",
                "operationId": "23d6b27133a71f392239e75377272c27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "subject"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Confirmation inscription"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "transactional"
                                    },
                                    "target_types": {
                                        "type": "string",
                                        "example": "JOUEURS"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Bienvenue sur EcoloQuiz"
                                    },
                                    "sender_name": {
                                        "type": "string"
                                    },
                                    "sender_email": {
                                        "type": "string"
                                    },
                                    "body_html": {
                                        "type": "string"
                                    },
                                    "body_text": {
                                        "type": "string"
                                    },
                                    "variables_schema": {
                                        "type": "object"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Template created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/email-templates/{id}": {
            "get": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Get a specific email template",
                "operationId": "2b166a5c0b88581438f954c14801c123",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Template ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template details"
                    },
                    "404": {
                        "description": "Template not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Update an email template",
                "operationId": "d98118f4ea4f3074a80329dac5f8de08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Template ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "target_types": {
                                        "type": "string"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "sender_name": {
                                        "type": "string"
                                    },
                                    "sender_email": {
                                        "type": "string"
                                    },
                                    "body_html": {
                                        "type": "string"
                                    },
                                    "body_text": {
                                        "type": "string"
                                    },
                                    "variables_schema": {
                                        "type": "object"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Template updated successfully"
                    },
                    "404": {
                        "description": "Template not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Email Templates"
                ],
                "summary": "Delete an email template",
                "operationId": "d01ee3b98371ab624d81694dbe7b9bb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Template ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template deleted successfully"
                    },
                    "404": {
                        "description": "Template not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}